requirements.txt
requests==2.31.0
PyYAML==6.0.1
apscheduler==3.10.1
openapi-spec-validator==0.6.0
deepdiff==6.3.1
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0 